artemis/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Blake Johnson <blakeejohnson39@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.18.2"
fastapi = "^0.78.0"
python-multipart = "^0.0.5"
email-validator = "^1.2.1"
requests = "^2.28.0"
celery = "^5.2.7"
passlib = "^1.7.4"
tenacity = "^8.0.1"
pydantic = "^1.9.1"
emails = "^0.6"
raven = "^6.10.0"
sentry-sdk = "^1.5.12"
gunicorn = "^20.1.0"
Jinja2 = "^3.1.2"
alembic = "^1.8.0"
SQLAlchemy = "^1.4.37"
python-jose = "^3.3.0"
redis = "^4.2.2"
motor = "^3.0.0"
odmantic = "^0.5.0"
redis-om = "^0.0.27"
sqlmodel = "^0.0.6"
asyncpg = "^0.26.0"
PyYAML = "^6.0"
beanie = "^1.15.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pre-commit = "^2.19.0"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.961"
commitizen = "^2.28.0"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.5.0"
hypothesis = "^6.49.0"
hypothesis-jsonschema = "^0.22.0"
sqlalchemy-stubs = "^0.4"
Flake8-pyproject = "^0.9.0"
pytest-asyncio = "^0.19.0"
httpx = "^0.23.0"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
line_length = 88
force_grid_wrap = 0
include_trailing_comma = true
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
